backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Azure OpenAI
openai==1.10.0
azure-identity==1.15.0

# Vector Database
chromadb==0.4.22
azure-search-documents==11.4.0

# RAG Framework
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.3

# Evaluation
ragas==0.1.1
datasets==2.16.1

# Excel Generation
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Logging & Monitoring
loguru==0.7.2
